Refine project structure, Automate build & testing
Issues resolved:

 - The previous project structure made it difficult to identify the
different units of logical concern in the project. This would have made
it difficult for new contributors to grasp the inner workings of the
project and locate files relevant to features to be implemented.

    This patch refactors the modules into appropriately named and
    located packages as well as changing all import statements to
    reflect this change.

 - The project lacked a setup.py and  did not conform to the
prevalent approach to project structure in the python ecosystem.
This prevented the project from being able to be built and intalled
using pythons setuptools, as is customary, and made it incompatable
with the PyPi repository for distribution.

    This patch introduces an appropriate setup.py file and refactores
    the location of source files to match prevailing python conventions

Automation:

Developers contributing to this project are required to run their
code through a myriad of linting and testing tools to ensure that
their contributions are working and do not violate any of the
projects conventions. Users and developers alike were required to
install a list of dependencies manually before begining to work
with the project. This workflow was cumbersome and would promote
skipping these vital steps in the development cycle.

This patch introduces a makefile as a central location for the
automated running of tests and lints as well as the installation
of all required dependencies through the make utility.

 - The previous instaltion process required the user to manually
 install a list of dependencies before being able to test or run
 the project. It also had developers install the dependencies
 globally, which may cause version conflicts with their existing
 installations for other projects.

    To solve this, this patch's makefile includes a list of its
    dependencies which can be installed with a single command.
    (make setup or make dependencies).
    The makefile also sets up a virtual environment for these
    installations so as to not conflict with existing package
    installations the contributor might have.

 - The previous testing process involved a python script to run
 a list of executable python tests. This approach may have been
 unfamiliar to developers and does not offer much diagnostic
 information outside of whether or not the tests passed.

    This path introduces the nose test framework as a dependency
    and adds a test command to the makefile which will automatically
    detect and run proper tests anywhere in the project and offer
    a myriad of diagnostic information while doing so.